+++ database/migrations/create_content_tags_table.php
@@ -22,6 +22,7 @@ return new class extends Migration
             $table->text('abstract')->nullable();
             $table->longText('description')->nullable();
             $table->json('metadata')->nullable();
+            $table->json('images')->nullable();
             $table->boolean('active')->default(true);
             $table->boolean('show_in_menu')->default(false);
             $table->boolean('in_evidence')->default(false);