+++ src/ContentoServiceProvider.php
@@ -26,6 +26,7 @@ class ContentoServiceProvider extends PackageServiceProvider
                 'create_mail_forms_table',
                 'create_modals_table',
                 'create_content_tags_table',
+                'update_content_tags_add_images',
                 'create_content_taggables_table',
                 'create_settings_table',
             ])